% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{curator}
\alias{curator}
\title{curator}
\usage{
curator()
}
\description{
This function calls the shiny app, allowing users to input and save metadata
and generate DDI Codebooks and README files.
}
\examples{
curator()

}
