% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIYDDICurator-package.R
\docType{package}
\name{diyddi-package}
\alias{diyddi}
\alias{diyddi-package}
\title{diyddi: A Curation Tool for Human Survey Research Data Based on the rddi Package}
\description{
The package serves as a graphic user interface to the rddi package and allows for researchers and curators working with human survey data to gather appropriate metadata to generate a DDI 2.5 Codebook.
}
\author{
\strong{Maintainer}: Daniel Woulfin \email{dw2896@nyu.edu} (\href{https://orcid.org/0000-0001-5502-9861}{ORCID})

Other contributors:
\itemize{
  \item Global TIES for Children (https://steinhardt.nyu.edu/ihdsc/global-ties) [copyright holder]
}

}
\keyword{internal}
